Fix decorator test to work with the new expansion logic
Before the refactoring of the core logic, multiple decorators macro
were executed top-down, from the outer to the inner. Now this logic is
reversed because it seems more right to me so now in this test code:

@outer
@my_macro2
@middle
@my_macro
@inner
def run():
    x = 10
    x = x + 1
    return x

``my_macro`` is expanded **before** ``my_macro2``, and ``my_macro2``
receives the expanded tree as parameter.