Allow "SELECT * FROM foo" emulation in pinq. Fixes #2
This is done by passing both the table and its columns
collection (before this patch only the latter was passed) to the
function that executes the 'select()' the former gets a name with
'_alias' added to it.

Then each occurrence of the target name ('foo' in 'foo for foo in
bar') is checked and if not is further restricted to one of its
attributes (so there no 'foo.zoo'), the names is changed in
'foo_alias'.